Fix utils.rs duplicate function and unused imports after index split
